+++ adam/entities/__init__.py
@@ -1,2 +1,3 @@
 from ..core.entities import Configuration
 from ..features.simulation.entities import Data, Collision, CollisionData, ConfigurationData
+from ..features.map_maker.entities import Body, Box, Capsule, Cube, Cylinder, Sphere