+++ adam/features/map_maker/__init__.py
@@ -1,0 +1,1 @@
+from .map_maker import MapMaker