+++ adam/features/map_maker/entities/__init__.py
@@ -1,0 +1,6 @@
+from .body import Body
+from .cube import Cube
+from .cylinder import Cylinder
+from .sphere import Sphere
+from .capsule import Capsule
+from .box import Box