+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.2.0
+- Added a whole new feature where bodies can be added to the scene by code.
+- Added a method to use the MuJoCo engine to just check the collisions on the given configuration list
+
 # 0.1.1
 - Added a default key-value pair to the collision dictionary
 - Added a method for the ```Simulation``` class where the collision dictionary can be extended